error handling added

+++ web3/blockchain/createTransaction.go
@@ -22,7 +22,7 @@ func signTransaction(privateKey *ecdsa.PrivateKey, user *User) (*big.Int, *big.I
 	return r, s, nil
 }
 
-func createTransaction(from *ecdsa.PublicKey, to *ecdsa.PublicKey, amount int, R big.Int, S big.Int) *Transaction {
+func createTransaction(from *ecdsa.PublicKey, to *ecdsa.PublicKey, amount float64, R big.Int, S big.Int) *Transaction {
 	return &Transaction{
 		From:   from,
 		To:     to,
